More cleaning up code.
