+++ .gitignore
@@ -37,3 +37,6 @@ model/economy/trade/__pycache__/__init__.cpython-311.pyc
 model/economy/trade/__pycache__/market.cpython-311.pyc
 model/population/__pycache__/group_properties.cpython-311.pyc
 model/economy/industry/__pycache__/firm_properties.cpython-311.pyc
+/.vscode
+/.continue
+model/province/__pycache__/province_properties.cpython-311.pyc
